4.3.57 Исправление ошибок Адаптация под PHP7 Восстановил кэширование эскизов Поправил кэширование видимых дочерних страниц в PageRecursiveCache Яндекс.Дзен в SocialProfile Исправил мобильное меню в андроиде Исправил, чтобы форма поиска не выдавала подсказки после отправки Статические значения в viewportprops AJAXForm переписал на API Условный favicon (если присутствует) Отказался от wrapFields в пользу Vue-атрибутов Модальное окно обратной связи и заказа звонка сделал через Vue Мобильное меню сделал через API Сделал поддержку FontAwesome 6

+++ classes/semantic/snippet.class.php
@@ -104,7 +104,11 @@ class Snippet extends SOME
         }
         $this->modify_date = $datetime;
         $this->editor_id = $uid;
-        if ($this->id && ($this->updates['urn'] ?? false) && ($this->updates['urn'] != $this->properties['urn'])) {
+        if ($this->id &&
+            ($this->updates['urn'] ?? false) &&
+            ($this->properties['urn'] ?? false) &&
+            ($this->updates['urn'] != $this->properties['urn'])
+        ) {
             $this->deleteFile();
         }
         parent::commit();
